make fancier setup

+++ util/chroot-setup.sh
@@ -1,12 +1,12 @@
 #!/bin/sh
 
-mount proc -t proc /proc
-mount -t sysfs sys sys
+[ -d /proc/1 ] || mount proc -t proc /proc
+[ -d /sys/class ] || mount -t sysfs sys /sys
 [ -c /dev/urandom ] || mdev -s
 
-ln -s /proc/self/fd /dev/fd
-ln -s /dev/fd/0 /dev/stdin
-ln -s /dev/fd/1 /dev/stdout
-ln -s /dev/fd/2 /dev/stderr
+[ -d /proc/self/fd ] && ln -s /proc/self/fd /dev/fd
+[ -L /dev/fd/0 ] && ln -s /dev/fd/0 /dev/stdin
+[ -L /dev/fd/1 ] && ln -s /dev/fd/1 /dev/stdout
+[ -L /dev/fd/2 ] && ln -s /dev/fd/2 /dev/stderr
 
 exec /bin/bash -li
